Skrócenie komentarzy

+++ 01_import_and_fixes.sql
@@ -6,6 +6,6 @@ ALTER COLUMN kwota FLOAT;
 ALTER TABLE TabelaTransakcje
 ALTER COLUMN data_transakcji DATETIME2;
 
--- 3. Standaryzacja klucza obcego (konto_id) na INT
+-- 3. Standaryzacja klucza obcego (konto_id) na SMALLINT
 ALTER TABLE TabelaTransakcje
-ALTER COLUMN konto_id INT;
+ALTER COLUMN konto_id SMALLINT;
